Create a Bucket class
Rather than having a loose function for this, we can just subclass the
`pulumi_aws.s3.Bucket` class directly.

It also pulls it out of the `util` module, which makes it a bit more
discoverable.
